requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
playwright>=1.40.0
solvecaptcha-python>=1.0.0
